Second look at item queueing
Gets items from the server queue into our own queue, and tries to add from that one.
Requires a "MatchListItem" condition which takes an int and returns if it was the index of the last item the user received.
If it fails, it will try to receive again in 5 seconds.